+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 ### Adicionado
 - Estrutura do HTML.
 - Estilização do CSS.
+
+## [Versão 0.0.8] - 2025-02-26
+### Adicionado
+- JavaScript.
+- Funcionamento da tabela.